Remove deprecated Nav component and update global styles
