--- src/app/ui/nav/Heading.module.css
@@ -1,13 +1,0 @@
-.list {
-  display: flex;
-  list-style: none;
-  align-items: center;
-  gap: var(--size-3);
-
-  a,
-  a:hover {
-    color: var(--secondary-color);
-    text-decoration: none;
-    cursor: pointer;
-  }
-}
