+++ src/pages/about.md
@@ -1,5 +1,5 @@
 ---
-title: About Me
+title: About
 ---
 
 # About me
