+++ src/pages/index.md
@@ -4,7 +4,7 @@ title: MDwritr
 
 ![MDwritr logo](/mdwritr-logo.svg)
 
-# Welcome to MDwritr
+# Local markdown blogging
 
 MDwritr is a self-hosted markdown blogging platform powered by [NextJS](https://nextjs.org), [Open-Props](https://open-props.style/), [Marked-React](https://github.com/sibiraj-s/marked-react), and [Ariakit](https://ariakit.org/). Edit `/src/pages/index.md` to make changes to this page or add your first markdown post to `/src/posts`.
